Fixed jest improved security and removed sequelize

+++ server/__tests__/utils/truncate.js
@@ -1,10 +1,13 @@
 const models = require("../models");
+const excludes = ["knex"];
 
 module.exports = () => {
   return Promise.all(
     Object.keys(models).map(key => {
-      const model = models[key];
-      return model.query().truncate();
+      if (!excludes.includes(key)) {
+        const model = models[key];
+        return model.query().truncate();
+      }
     })
   );
 };
